Disable default user password and basic auth
Ensured that the default user is created with a random username and
password. This prevents the default message from being logged at boot
time. The framework creates the user always, so we cannot turn off
its instantiation. However, we explicitly test that basic auth is
disabled, and that is the only purpose for the default user.

[#129582609] Default users and Spring basic auth should be disabled

Signed-off-by: Sebastian Wittenkamp <swittenkamp@pivotal.io>