Use binary types for UUIDs
This reduces the space required for UUIDs and should improve indexing
speed once they are used as as index. By making the conversion now,
rather than after we start using UUIDs as primary keys and potentially
see a big performance hit, we hopefully avoid a painful migration in the
future.

[#131746317] Make UUIDs the primary key for secrets in the database