set secret via path or name in request body
The test refactoring in here isn't the prettiest but it seemed like a
pragmatic solution while we have to maintain backwards compatibility. I
didn't copy all tests to have both a "via path" and "via body" version.
I reasoned that when we implement "via body only" the existing tests
that cover both cases will serve as a reference. YMMV.

[#132716063] Credential name should be a request parameter of update request