Bail if mysql fails to start
Periodically, MySQL will fail to start. The service command appears to
still exit 0 when this happens. This means that when we try to create
the test db, it will hang for a long time before finally timing out.

With this change, we check status explicitly via the mysqladmin tool and
blow up if we failed to connect. This should help the pipeline fail fast
if there's an issue.

Also, supply the --no-daemon flag to gradle since we don't ever need to
start the daemon in the pipeline. This should help shave a little bit of
time off the test run.

[#133943771] Unit tests intermittently run forever