Implement the DSM encryption provider
Note that the DSM only supports AES-CCM mode, not AES-GCM, so we made
the AES cipher dependant on the current EncryptionConfiguration.

We made the nonce/IV length 12 bytes instead of 16 based on the example
in the Dyadic Programmer Guide. 16 might also work - both in terms of
the code and cryptographic security.

[#133697493] User should be able to use Dyadic DSM as encryption provider

Signed-off-by: Jen Hamon <jhamon@pivotal.io>