Fix canary value
Previously, we generated the canary value's length based on the
encryption key. However, we started hard-coding it to support the DSM,
which does not support key encoding, such that we couldn't check the
length.

The initial hard-coded size (256 bytes) is different from the size of
existing, deployed canaries (believed to be 128 bytes for all deployed
instances). We are therefore switching to 128 bytes for backwards
compatibility.

[#135791143] add recent credhub to super-bosh to fix the pipeline

Signed-off-by: Phil Goodwin <pgoodwin@pivotal.io>