Revert "Clean up data service tests"
While this fix worked for MySQL and H2 UUIDs (stores as a byte[]), it failed to correctly handle
Postgres UUIDs (stored as a native UUID).

The fix for Postgres as shown at
http://crafted-software.blogspot.com/2013/03/uuid-values-from-jdbc-to-postgres.html
doesn't work with the byte[] UUIDs, and managing the two types in these
tests would end up being messier than it's worth.

This reverts commit 15531bd983b937989a8775cae46b01cf1d556d8f.