Encrypt and decrypt based on the relevant key
At startup, we generate a map of canaries to keys.

We omit any keys that are not active and do not already have a
matching canary in the database. This is intended to help prevent and
clean up superfluous canaries.

NOTE: We can no longer assume that the database has an issue if a canary
can't be decrpyted. This presents a new failure case since no secrets
will be decryptable if the matching encryption key's canary is in an
error state. Database backups are therefore that much more important.

[#133889481] User should be able to rotate dev_internal keys