Retry encryption operations when Luna disconnects
Created a wrapper class for Ciphers (which have final methods) so that
we can mock them out in tests.
Changed the LunaEncryptionService constructors so that mocks can be
injected into it during testing.
Had to write pretty convoluted code to get this functionality tested,
not convinced that the value add was worth it.

[/#133565251] Update version to 0.4.0 for release

Signed-off-by: Phil Goodwin <pgoodwin@pivotal.io>