Always encrypt secrets, even if value is unchanged
The logic around preventing double encryption added a lot of complexity,
as well as the diminished performance of an extra resource-intensive
operation, in order to add a fairly small optimization (i.e., to try to
minimize overuse of an HSM key).

It may make sense to add this logic back in if and when we add more
sophisticated measures for handling HSM key overuse as part of a future
story.

[#133876467] CredHub should gradually rotate data to new encryption key while it is running

Signed-off-by: Isobel Redelmeier <iredelmeier@pivotal.io>