Consolidate password encryption
Password parameters now borrow the encryption key UUID used for the
password itself, rather than keeping track of a separate encryption key
UUID. While this forces us to keep closer control of when the password's
encryption key UUID gets changed - lest we become unable to
decrypt/rotate the parameters - it keeps the data model simpler.

Also improves SecretEncryptionHelper.rotate so that the plaintext is
decrypted zero or one times, whereas it used to be decrypted twice.

[#137188079] User should be able to rotate DSM encryption key

Signed-off-by: Sebastian Wittenkamp <swittenkamp@pivotal.io>