Move retry logic to its own class
Introduced RetryingEncryptionService to encapsulate all the retry code
including locking.
SecretEncryptionHelper depends on RetryingEncryptionService but
EncryptionKeyCanaryMapper does not. This prevents stack overflows when
decryption fails.
EncryptionKeyCanaryMapper initializes the key mappings when it is
created.
Introduced RemoteEncryptionConnectable as an interface implemented by
LunaConnection, DyadicConnection and the new BCNullConnection. It is the
way the RetryEncryptionService asks for a reconnect before retrying the
encryption operation.

[#136539551] CredHub should work appropriately with HA HSM

Signed-off-by: Phil Goodwin <pgoodwin@pivotal.io>