Unify validations, pull up "type"
Entangled in a single commit to ensure we reach the finish line safely.

Essentially, this commit does two things. The first is to introduce a
unifying compromise for validations. We retain the ability to use JSR303
validation annotations, but these are triggered manually in the
SecretsController, instead of by Spring's @Validated annotation. This
leaves open using either simple JSR303 annotations with an escape hatch
to more complex validations. Ideally, in a saner world, we will be able
to hoist the latter into Validator objects.

The second change introduced is to pull up "type" into
BaseSecretRequest from duplicate implementations in BaseSecretSetRequest
and BaseSecretGenerateRequest.

Remaining changes are to consolidate error messages.

[#140582047] Use Jackson for password generate requests

Signed-off-by: Andrew Cantino <acantino@pivotal.io>