Reject MTLS if missing client_authentication key
We think this was a latent bug.

Previous behavior was that if the certificate did not contain the
client_authentication key, then Spring Security would move on to the
Oauth2 filter and fail there.

This commit configures Spring Security's X509AuthenticationFilter to
terminate processing upon seeing an X509 authentication failure, rather
than noting the X509 failure and continuing to OAuth2.

In addition, we needed to change the error message in the thrown exception
and update X509AuthenticationFailureHandler.

[#138284147] client_auth extended key usage extension should be validated for MTLS authentication

Signed-off-by: Kelly Gerritz <kgerritz@pivotal.io>