JUnitize EncryptionKeyCanaryDataServiceTest
Avoids test pollution problem that was incorrectly handled by deleting
the repo before tests.

Firstly, @DataJpaTest only loads a subset of the app. Since we don't
read the application-test.yml, we don't start with a canary value in the
database. This avoids having to code around it.

Secondly, it makes all @Test methods transactional by default. So we
don't need to clean up either before or after.

Bonus: It's faster. Much faster in IntelliJ.

- Spectrum -> SpringRunner
- @SpringBootTest -> @DataJpaTest
- Build subject in @Before method
- Replace describe/it blocks with @Test methods
- Use hasSize() instead of .size(), equalTo()

[#143069351] Fix failing unit tests when running in parallel with gradle

Signed-off-by: Jacques Chester <jchester@pivotal.io>