EncryptionKeyCanaryDataServiceTest runs on all DBs
Previously we adjusted this test to only run on H2, as it was failing on
MySQL and PosgreSQL. It was failing on those two DBs because of state
left by other tests -- canaries that were inserted as a side-effect of
launching the app via @SpringBootTest.

This commit reintroduces a workaround we use universally, which is to
delete records in the database before proceeding to test. Our intention
is to remove all of these simultaneously when the tests have been
JUnitized.

- Add @AutoConfigureTestData config to prevent H2 from being used in all
  tests.
- Remove @ContextConfiguration
- Reintroduce DatabaseProfileResolver

[#143069351] Fix failing unit tests when running in parallel with gradle

Signed-off-by: Jacques Chester <jchester@pivotal.io>