Generate User type at api level
Adds a new user type to the generate requests and reuses the password
generation parameters as a lot of these values are similar. The User
generator simply calls to the existing password generator as well.

UserGenerationParameters always constructs the same two
passwordGenerationParameters so only these default values can be used.

[#143515813] Users should be able to generate 'user' type via API