Reduce unit-test-postgres connection pool size
After the previous commit, we saw database connection errors related to
the number of users. Preliminary investigation shows that the connection
pool is not shared across all tests; instead there seems to be a
connection pool per-context.

This commit brings the connection pool settings down to the minimum size
that appears to work.

[#144592583] Release 0.7.0 reports version on /info as 'dev'

Signed-off-by: Jacques Chester <jchester@pivotal.io>