Reintroduce a 'placeholder' version
Removing info.app.version works locally, where we run in the 'dev'
profile. And it works in unit tests, where we run in various 'unit-test'
profiles.

It doesn't work if you don't provide a profile. If, for example, you are
a BOSH-templated application.yml that doesn't yet understand the concept of
versions and names.

This change reintroduces a version, intended to keep things working
until we can update the BOSH release to correctly write this information
through application.yml.erb.

[#144592583] Release 0.7.0 reports version on /info as 'dev'

Signed-off-by: Edie Beer <ebeer@pivotal.io>