fix: CredHub 2.7.0 can be upgraded
* We needed to downgrade the FlyWay version from 6 to 5 in commit
803cd285df9e7440dc52832c05ea17060877b048 due to another
[issue](https://www.pivotaltracker.com/story/show/173841985)

* This caused an installation error on upgrade from CredHub 2.7.0 to
2.7.1 reporting a migration type mismatch. This is because Credhub 2.7.0
contains FlyWay 6 which includes a database migration incompatible with
FlyWay 5 (some types in the Java code changed, which forced the values
in the "type" column of the migration table to be converted from
"SPRING_JDBC" to "JDBC").

* This requires the database to be repaired, bringing the migration
types in the DB back in sync with the declared types. This is what this
commit does.

* We can remove this code should we upgrade to FlyWay 6 again in the
future, since FlyWay 6 is able to handle these situations without
special instruction.

* We cannot upgrade to FlyWay 6 until we can validate that we no longer
need compatability with older versions of MariaDB. Current compatability
matrix for FlyWay and MariaDB is
[here](https://flywaydb.org/documentation/database/mariadb).

* We tested this manually by deploying a CredHub 2.7.0 to get the state
we wanted. We deployed CredHub 2.7.1 and saw the upgrade fail. We then
deployed with our changes and the upgrade was succesful. We don't see
much room for unit testing a change like this and believe the best
coverage would be a pipeline job. We don't currently have a setup to add
a test like this to our pipelines. We should reconsider upgrade
pipelines that cover more upgrade scenarios.

[#173911070](https://www.pivotaltracker.com/story/show/173911070)

Co-authored-by: Bruce Ricard <bricard@vmware.com>