Remove usages of deprecated media type for linting
The APPLICATION_JSON_UTF8 media type was deprecated in Spring 5.2 in
favor of just using APPLICATION_JSON and setting a character encoding
when used as a contentType. The linting script complains about these
usages.

It is unclear what the intention behind using APPLICATION_JSON_UTF8 was,
but testing character encoding feels out-of-scope for tests that are
being used to generate documentation.

Authored-by: Brian Upton <bupton@vmware.com>
Authored-by: Pablo Rodas <prodas@vmware.com>