Make duration overridden non-nullable
Removes the null case from all usages of the new durationOverridden
field. Because the existing CertificateView is used for all
certificate-related endpoints (including GET, DELETE, etc), a new view
class had to be introduced to prevent the duration overridden field from
showing up on inappropriate endpoints.

The new CertificateGenerationView inherits from the original
CertificateView and only adds fields that are specific to the generate
and regenerate endpoints.

Co-authored-by: Brian Upton <bupton@vmware.com>
Co-authored-by: Pablo Rodas <prodas@vmware.com>
Co-authored-by: Preethi Varambally <pvarambally@pivotal.io>