re-bump: bc-fips from 1.0.1 to 1.0.2.3
* Combines a previous commit for the bump plus another commit with necessary test changes because the test change depends on the bump.
* Also fixes two backend tests that failed in CI.
* Previously, we rely on bc-fips-provider lib to choose a default
SecureRandom when it generates keys. This default SecureRandom
implementation is not guaranteed to be be backed by a "nonblocking"
entropy source, the recommended mode for web app (a "blocking"
entropy source will "block" while the system gathers more entropy,
thus creating performance issues). See more details in the story.
* In the current version of bc-fips-provider, the default SecureRandom
that is chosen does not have observable performance issue in
production or in tests. However, the next version of bc-fips-provider
(1.0.2.3) will change the default to a "blocking" entropy source,
which will potentially create performance issue for production and
tests. We have validated that a "blocking" entropy source will slow
down some tests by orders of magnitude, resulting in test timeouts in
CI. Though we have not been able to observe performance issues in
production, we could not rule out the possibility.
* Hence, to ensure that a non-blocking entropy source (see story for
more research on why this mode is both secure and performant)
regardless of which bc-fips-provider version is used, we configure the
provider to use `NativePRNGBlocking` SecureRandom by default in both
production and in tests.

Co-authored-by: Hongchol Sinn <hsinn@vmware.com>
Co-Authored-By: Peter Chen peterch@vmware.com

[#181856904]