DefaultCertificateService.findAllMetadata: retrieve data in batches
This fixes an issue with Postgres and instances with >= 32768 certificates
available to a user.

CertificateDataService.findAllValidMetadata(names) puts all cert names in a
single IN (...) SQL query, which JdbcTemplate turns into a query with a bind
variable for each cert name (IN (?, ?, ?, ...)). This exhausts Postgres's
per-query bind variable limit.

Fixes #340