Revert "fix: error when deleting certificate version"
This reverts commit 830bb16eade21e98cc74b6916419f98fa279a719.

This commit (along with other commits in
https://github.com/cloudfoundry/credhub/pull/711/commits) is causing
an issue where CredHub might fail to start if CredHub is configured with
a database vendor that enforces additional access control on database
triggers creation (e.g. Amazon RDS), as this approach will attempt to
create database triggers at start time. See more context in: #187367323
and #187373461