Revert "fix: Orphaned encrypted_value records for H2"
This reverts commit adac39b124f517e8a2bc7c062ca522dc35649ae4.

This commit (along with other commits in
https://github.com/cloudfoundry/credhub/pull/711/commits) is causing
an issue where CredHub might fail to start if CredHub is configured with
a database vendor that enforces additional access control on database
triggers creation (e.g. Amazon RDS), as this approach will attempt to
create database triggers at start time. See more context in: #187367323
and #187373461