Revert "fix: Orphaned encrypted_value records"
This reverts commit 0e2cef67bfd1b7f12c5f9ea69c1448c89e0b2c86.

This commit (along with other commits in
https://github.com/cloudfoundry/credhub/pull/711/commits) is causing
an issue where CredHub might fail to start if CredHub is configured with
a database vendor that enforces additional access control on database
triggers creation (e.g. Amazon RDS), as this approach will attempt to
create database triggers at start time. See more context in: #187367323
and #187373461