fix: api docs build failure
- problem: script `/scripts/generate_documentation_snippets.sh`
fails with:
```
In plugin 'org.asciidoctor.convert' type 'org.asciidoctor.gradle.AsciidoctorTask' property 'logDocuments' is annotated with @Optional but that is not allowed for 'Console' properties.
```

- It looks like our asciidoctor lib is out of date and not compatible
with the new gradle. Hence, taking advice from
https://github.com/asciidoctor/asciidoctor-gradle-plugin/issues/600,
migrate to a supported lib instead (`org.asciidoctor.jvm.convert`);
and use the latest version of that.

- The changes in this commit follows this sample app: https://github.com/spring-projects/spring-restdocs-samples/blob/c98c705ad9043aa19db96ce29160e0fbdfeaf8bc/restful-notes-spring-hateoas/build.gradle

- note: The directory changes in `credhub/backends/credhub/src/docs/asciidoc/index.adoc`
is required as part of this migration/upgrade; without these changes,
the rendered doc html will not correctly import those "snippets". The
new lib also seems to output the resulting html to a different dir,
so update tasks and readme to reflect that.

[#187494474]