fix: migration error from H2 bump
- as a result of H2 bump from v1->v2, there is now
migration error associated with the creation of
these constrainsts
- but since these constraints will be dropped in the
later migrations anyway, so in this commit, we simply
remove both the creation and the dropping of these
constrainsts.
- the migration history of h2 doesn't really matter,
as it is just test database whose main purpose is to
be as close to mysql/postgres as possible in terms of
the final schema.

[https://vmw-jira.broadcom.net/browse/TNZ-23977]