temp fix: strange h2 column type mismatch error
- the error:
```
org.hibernate.tool.schema.spi.SchemaManagementException: Schema-validation: wrong column type encountered in column [credential_uuid] in table [credential_version]; found [binary (Types#VARBINARY)], but expecting [varbinary (Types#BINARY)]
``
Which is strange because it is unclear what "expecting [varbinary (Types#BINARY)]"
means. Also, this does not match up with the set up in MySQL
where this column's datatype is VARBINARY, which is working fine.

- it is unclear why the change in this commit (changing VARBINARY into BINARY)
fixes the problem, even though it creates further divergence from MySQL
(which is not good because H2, as a test DB, is mainly here to
simulate real prod DBs like MySQL and Postgres); hence, this fix
should be treated as a temp fix that should be revisited

- we should revisit this problem, once we bump to spring boot 3 (and
bumping other related DB libs in the process, such as Hibernate, Jakarta
Persistence API, etc.); as it is possible that when all these
inter-connected libs are all bumped, this H2 problem will be more
solve-able.

[https://vmw-jira.broadcom.net/browse/TNZ-23977]