API doc improvement
- edit for clarity
- the original language:
```
* If the duration used to generate the currently active version of the certificate, or the duration provided via the respective field, is lower than the minimum duration, the regenerated certificate will use the minimum duration instead and the response will contain the `duration_overridden` flag set to true. The duration value used to regenerate the certificate is included in the `duration_used` field of the response.
```
the logic of "if A, or B, then C" where:
A = the duration used to generate the currently active version of the certificate...is lower than the minimum duration
B = the duration provided via the respective field...is lower than the minimum duration
C = the regenerated certificate will use the minimum duration instead
is technically incorrect. For example, if A involves duration of 100
days, B 300 days, and the minimum duration is 200 days, then 300 days
will eventually apply, not 200 days.

* simplify the language